separate skia's part from build.rs, introduce a build configuration that is used to inform cargo for full builds and builds that use prebuilt binaries
